split meal location maps into views and presenter
