+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
         android:theme="@style/Theme.AyoEats"
         tools:targetApi="31">
         <activity
-            android:name=".activities.MealLocationMapsActivity"
+            android:name=".views.meallocationmaps.MealLocationMapsView"
             android:exported="false"
             android:label="@string/title_activity_meal_location_maps"
             android:theme="@style/Theme.AyoEats" />
