+++ app/src/main/res/layout/activity_meal_location_maps.xml
@@ -5,7 +5,7 @@
     android:id="@+id/coordinatorLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.MealLocationMapsActivity">
+    tools:context=".views.meallocationmaps.MealLocationMapsView">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/appBarLayout2"
